+++ src/composables/spaceDetector.ts
