+++ src/plugins/fullscreen/store.ts
@@ -10,8 +10,8 @@ import type { FullscreenPluginOptions } from './types'
 import { acceptHMRUpdate, defineStore } from 'pinia'
 import { computed, ref } from 'vue'
 
+import { spaceDetector } from '@/composables/spaceDetector'
 import { useCoreStore } from '@/core/stores'
-import { spaceDetector } from '@/lib/spaceDetector'
 
 import { PluginId } from './types'
