+++ src/plugins/initialView/store.ts
@@ -3,8 +3,8 @@ import type { InitialViewPluginOptions } from './types'
 import { acceptHMRUpdate, defineStore } from 'pinia'
 import { computed } from 'vue'
 
+import { spaceDetector } from '@/composables/spaceDetector'
 import { useCoreStore } from '@/core/stores'
-import { spaceDetector } from '@/lib/spaceDetector'
 
 import { PluginId } from './types'
